app/requirements.txt
flask
groq
langchain-community>=0.0.1
python-dotenv
langdetect